+++ rest_auth/utils.py
@@ -1,9 +1,5 @@
 from six import string_types
-import sys
-if sys.version_info < (2, 7):
-    from django.utils.importlib import import_module
-else:
-    from importlib import import_module
+from importlib import import_module
 
 
 def import_callable(path_or_callable):
@@ -13,3 +9,8 @@ def import_callable(path_or_callable):
         assert isinstance(path_or_callable, string_types)
         package, attr = path_or_callable.rsplit('.', 1)
         return getattr(import_module(package), attr)
+
+
+def default_create_token(token_model, user, serializer):
+    token, _ = token_model.objects.get_or_create(user=user)
+    return token